enemies now spawn in world and follow player Many minor tweaks
git-svn-id: https://svn.neumont.edu:8443/svn/cplusplus_anthony_corbin@72 c70e0d21-08ce-ef44-a38d-91c21494a941

+++ GameSolution/Game/TargetMouse.cpp
@@ -2,5 +2,5 @@
 #include "Core.h"
 
 Vector2D TargetMouse::getPos() {
-	return Vector2D(Core::Input::GetMouseX(),Core::Input::GetMouseY());
+	return Vector2D((float)Core::Input::GetMouseX(),(float)Core::Input::GetMouseY());
 }
